+++ mne/tests/test_viz.py
@@ -113,10 +113,10 @@ def test_plot_evoked():
     evoked.plot(hline=[1], units=dict(mag='femto foo'))
     evoked_delayed_ssp.plot(toggle_proj=True)
     evoked_delayed_ssp.apply_proj()
-    assert_raises(RuntimeError, evoked_delayed_ssp.plot, toggle_proj=True)
+    assert_raises(RuntimeError, evoked_delayed_ssp.plot, proj='interactive')
     evoked_delayed_ssp.info['projs'] = []
-    assert_raises(RuntimeError, evoked_delayed_ssp.plot, toggle_proj=True)
-    assert_raises(RuntimeError, evoked_delayed_ssp.plot, toggle_proj=True,
+    assert_raises(RuntimeError, evoked_delayed_ssp.plot, proj='interactive')
+    assert_raises(RuntimeError, evoked_delayed_ssp.plot, proj='interactive',
                   axes='foo')